requirements.txt
Pillow==9.2.0
pynput==1.7.6
PySide6==6.4.0.1
pytesseract==0.3.10
